make sure roundtrip execute on top level directory. move out shared functionality to a utility script
